sign compare

+++ source/world/entity/SynchedEntityData.cpp
@@ -64,7 +64,7 @@ bool SynchedEntityData::isDirty() const
 
 void SynchedEntityData::clear()
 {
-    for (int i = 0; i < m_itemsArray.size(); i++)
+    for (size_t i = 0; i < m_itemsArray.size(); i++)
     {
         DataItem* item = m_itemsArray[i];
         SAFE_DELETE(item);
@@ -102,7 +102,7 @@ void SynchedEntityData::packAll(IDataOutput& dos) const
 
 void SynchedEntityData::assignValues(const ItemsArray& items)
 {
-    for (int i = 0; i < items.size(); i++)
+    for (size_t i = 0; i < items.size(); i++)
     {
         DataItem* newItem = items[i];
         DataID itemId = newItem->getId();
@@ -202,7 +202,7 @@ void SynchedEntityData::_WriteDataItem(IDataOutput& dos, const DataItem& dataIte
 
 void SynchedEntityData::Pack(const ItemsArray& items, IDataOutput& dos)
 {
-    for (int i = 0; i < items.size(); i++)
+    for (size_t i = 0; i < items.size(); i++)
     {
         const DataItem* item = items[i];
         if (item)
